network type id cache
