+++ OwlTree/src/spawning/ProxyFactory.cs
@@ -1,15 +1,26 @@
-
-
 using System;
+using System.Linq;
 
 namespace OwlTree
 {
     /// <summary>
     /// Responsible for generating network object proxies that will handle RPC send & recv logic.
-    /// Proxies and proxy factories are generated by OwlTree compiler step.
+    /// Proxies and proxy factories are generated by OwlTree source generator.
     /// </summary>
     public abstract class ProxyFactory
     {
+        /// <summary>
+        /// Gets the specific project implementation
+        /// </summary>
+        public static ProxyFactory GetProjectImplementation()
+        {
+            return (ProxyFactory)Activator.CreateInstance(
+                AppDomain.CurrentDomain.GetAssemblies()
+                    .SelectMany(a => a.GetTypes())
+                    .Where(t => t.IsClass && !t.IsAbstract && typeof(ProxyFactory).IsAssignableFrom(t)).FirstOrDefault()
+                );
+        }
+
         /// <summary>
         /// Creates a new proxy that wraps the given NetworkObject instance.
         /// </summary>
@@ -29,33 +40,5 @@ namespace OwlTree
         /// Gets the network object type from the given id.
         /// </summary>
         public abstract Type TypeFromId(byte id);
-        
-        /// <summary>
-        /// Used to created the proxy factory generated for this app.
-        /// </summary>
-        public static Type AssignedProxyFactory = typeof(ProxyFactoryDefault);
-    }
-
-    internal class ProxyFactoryDefault : ProxyFactory
-    {
-        public override T CreateProxy<T>(T instance)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override bool HasTypeId(Type t)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override Type TypeFromId(byte id)
-        {
-            throw new NotImplementedException();
-        }
-
-        public override byte TypeId(Type t)
-        {
-            throw new NotImplementedException();
-        }
     }
 }
